Remove 'command' from ping for IP in notification
Removing 'command' from ping for IP address in notification as requested by @fniephaus.

It was initially added because in rare cases, some shell integrations (i.e ZSH plugins) override the default command for ping with a function. In my case, ping is actually a function that accepts and parses URLs in addition to domains (e.g. ping https://github.com). 

However, the fallacy would actually be on the plugin in this case, and it is unlikely that this would be loaded into the bash shell within Alfred anyways, so we're removing it.